[filegen] move NixpkgsInfo into filegen (no more plansdk)
